The disassembler cannot be trusted to get a correct representation
Since we have no idea of what is code and what is data when reading from
RAM we cannot be sure what is actually code and thus cannot disassemble
correctly and we cannot make sure that labels are assigned to the
correct address. The "Disassembly" view is disabled for now and should
be used to display contents of the .PRG file instead.

We must use the Kickassembler "-debugdump" option instead to find the
correct address for a source location and thus get the ability to manage
breakpoints directly in the source.